revamp GUI
