+++ src/main/java/com/napnie/tfec/ui/FareCalculatorUI.java
@@ -27,30 +27,6 @@ import com.napnie.tfec.ui.StepUI;
  */
 @SuppressWarnings("serial")
 public class FareCalculatorUI extends JFrame {
-	private FareCalculator estimator;
-	
-	/** Input for origin. */
-	private JTextField origin;
-	/** Input for destination. */
-	private JTextField destination;
-	/** Button for fare calcurate action. */
-	private JButton estimate;
-	
-	/** Input for start fare rate. */
-	private JTextField startFare;
-	/** Input for running fare rate. */
-	private JTextField runFare;
-	/** Input for waiting fare rate. */
-	private JTextField waitFare;
-	
-	/** TextField for show distance between origin and destination. */
-	private JTextField distance;
-	/** TextField for show duration from origin and destination. */
-	private JTextField duration;
-	/** TextField for show estimated wait time of traveling. */
-	private JTextField waitTime;
-	/** TextField for show estimated taxi fare from origin and destination. */
-	private JTextField fare;
 	
 	private JLabel status;
 	
@@ -59,23 +35,23 @@ public class FareCalculatorUI extends JFrame {
 	private RouteMapGUI centerPanel;
 	
 	public FareCalculatorUI(FareCalculator estimator) {
-		this.estimator = estimator;
+		initComponents(estimator);
 	}
 	
 	/** Run this GUi. */
 	public void run() {
 		setDefaultCloseOperation(EXIT_ON_CLOSE);
-		initComponents();
 		setVisible(true);
 	}
 	
 	/** Initialize Components in this GUI. */
-	private void initComponents() {
+	private void initComponents(FareCalculator estimator) {
 		setTitle("Taxi Fare Estimate Calculator");
 		
-		JPanel leftPanel = initilizeLeftPanel();
+		stepUI = new StepUI();
 		centerPanel = new RouteMapGUI();
-		JPanel rightPanel = initilizeRightPanel();
+		ResultPanel result = new ResultPanel();
+		InfoPanel info = new InfoPanel(result, centerPanel, estimator);
 		
 		status = new JLabel();
 		status.setBackground( Color.BLACK);
@@ -83,174 +59,20 @@ public class FareCalculatorUI extends JFrame {
 		status.setText("Idle");
 		
 		setLayout(new BorderLayout());
-		add(leftPanel, BorderLayout.WEST);
+		add(info, BorderLayout.WEST);
 		add(centerPanel, BorderLayout.CENTER);
-		add(rightPanel, BorderLayout.EAST);
-		add(status, BorderLayout.SOUTH);
+		add(result, BorderLayout.EAST);
+		add(status, BorderLayout.NORTH);
 		
-		startFare.setText( String.valueOf( estimator.getStartFare() ) );
-		runFare.setText( String.valueOf( estimator.getRunFare() ) );
-		waitFare.setText( String.valueOf( estimator.getWaitFare() ) );
-		
-		setAction();
+		info.setStartFare( estimator.getStartFare() );
+		info.setRunFare( estimator.getRunFare() );
+		info.setWaitFare( estimator.getWaitFare() );
 		
 		pack();
 	}
 	
-	private void setAction() {
-		ActionListener estimateAction = (event) -> { 
-//			new Thread( () -> estimateAction() ) ;
-			try {
-				estimateAction();
-			} catch(Exception e) {
-				status.setText(e.getMessage());
-			}
-		};
-		estimate.addActionListener( estimateAction );
-		origin.addActionListener( estimateAction );
-		destination.addActionListener( estimateAction );
-	}
-
-	/** Initialize right panel of this GUI. */
-	private JPanel initilizeRightPanel() {
-		JPanel rightPanel = new JPanel();
-//		rightPanel.setMinimumSize(new Dimension(300, 700));
-
-		stepUI = new StepUI();
-		
-		// for resultPanel
-		JPanel resultPanel = new JPanel();
-		resultPanel.setLayout(new BoxLayout(resultPanel, BoxLayout.Y_AXIS) );
-		resultPanel.add( new JLabel("Estimated Price") );
-		
-		JPanel result = new JPanel();
-		initilizeResult(10);
-		result.setLayout(new GridLayout(4, 3) );
-		initField(result, "Distance: ", distance, "km");
-		initField(result, "Duration: ", duration, "minute");
-		initField(result, "Estimated Wating Time: ", waitTime, "minute");
-		initField(result, "Overall Price", fare, "Baht");
-		
-		resultPanel.add(result);
-		
-		rightPanel.setLayout(new BorderLayout() );
-		rightPanel.add(stepUI, BorderLayout.NORTH);
-		rightPanel.add(resultPanel, BorderLayout.SOUTH);
-		
-		return rightPanel;
-	}
-	
-	/** Initialize TextField in Estimated Price panel. */
-	private void initilizeResult(int width) {
-		distance = new JTextField(width);
-		duration = new JTextField(width);
-		waitTime = new JTextField(width);
-		fare = new JTextField(width);
-		distance.setEditable(false);
-		duration.setEditable(false);
-		waitTime.setEditable(false);
-		fare.setEditable(false);
-		
-		distance.setHorizontalAlignment(SwingConstants.RIGHT);
-		duration.setHorizontalAlignment(SwingConstants.RIGHT);
-		waitTime.setHorizontalAlignment(SwingConstants.RIGHT);
-		fare.setHorizontalAlignment(SwingConstants.RIGHT);
-	}
-	
-	/** Initialize left panel of this GUI. */
-	private JPanel initilizeLeftPanel() {
-		JPanel leftPanel = new JPanel();
-		
-		JPanel place = new JPanel();
-		place.setLayout(new BoxLayout(place, BoxLayout.Y_AXIS) );
-		initilizePlaceInput(10);
-		estimate = new JButton("Estimate");
-		JLabel originText = new JLabel("Origin:");
-		JLabel destinationText = new JLabel("Destination:");
-		
-		place.add(originText);
-		place.add(origin);
-		place.add(destinationText);
-		place.add(destination);
-		place.add(estimate);
-		
-		JPanel fareRate = new JPanel();
-		fareRate.setLayout(new BoxLayout(fareRate, BoxLayout.Y_AXIS) );
-		fareRate.add(new JLabel("Fare Rate"));
-		initilizeFareSet(10);
-		
-		JPanel fareInput = new JPanel();
-		fareInput.setLayout(new GridLayout(3, 3) );
-		initField(fareInput, "Start: ", startFare ,"Baht");
-		initField(fareInput, "Fare per 1 km Run: ", runFare ,"Baht");
-		initField(fareInput, "Fare per 1 hour Wating: ", waitFare ,"Baht");
-		
-		fareRate.add(fareInput);
-		
-//		leftPanel.setLayout(new BoxLayout(leftPanel, BoxLayout.Y_AXIS) );
-		leftPanel.setLayout(new BorderLayout() );
-		leftPanel.add(place, BorderLayout.NORTH);
-		leftPanel.add(fareRate, BorderLayout.SOUTH);
-		leftPanel.setPreferredSize(new Dimension(410 , 450+75) );
-		place.setPreferredSize(new Dimension(leftPanel.getWidth(), 150) );
-		fareRate.setPreferredSize(new Dimension(leftPanel.getWidth(), 150));
-		
-		
-		return leftPanel;
-	}
-	
-	private void estimateAction() {
-		status.setText("Calculating...");
-		
-		String origin = this.origin.getText();
-		String destination = this.destination.getText();
-		
-		if( !estimator.isRouteEstimated() ) status.setText( estimator.readHint( estimator.getHint() ) );
-		
-		estimator.estimateRoute(origin, destination);
-		if( !estimator.isRouteEstimated() ) return;
-		distance.setText( formatResult( estimator.getDistance() ) );
-		duration.setText( formatResult( estimator.getDuration() ) );
-		waitTime.setText( formatResult( estimator.getWaitTime() ) );
-		fare.setText( formatResult( estimator.estimateFare() ) );
-		
-		status.setText("Route Ploting");
-		Route route = estimator.getRoute();
-		centerPanel.setMap(route.getOriginLocation(), route.getDestinationLocation());
-		
-		stepUI.setStep(route);
-		
-		status.setText("Done");
-	}
-	
 	private String formatResult(double result) {
 		return String.format("%.2f", result );
-	}
-
-	/** Initialize input of Fare Rate panel. */
-	private void initilizeFareSet(int width) {
-		startFare = new JTextField(width);
-		runFare = new JTextField(width);
-		waitFare = new JTextField(width);
-	}
-	
-	/** Initialize input of Place panel. */
-	private void initilizePlaceInput(int width) {
-		origin = new JTextField(width);
-		destination = new JTextField(width);
-	}
-	
-	/**
-	 * Add TextField in panel
-	 * @param panel - JPanel that want to add TextField.
-	 * @param text - Text for describe TextField.
-	 * @param field - TexField to add in panel.
-	 * @param tailing - tailing unit of TextField.
-	 */
-	private void initField(JPanel panel ,String text, JTextField field, String tailing) {
-		panel.add(new JLabel(text) );
-		panel.add(field);
-		panel.add(new JLabel(" "+tailing) );
 	}
 	
 }
