+++ src/main/java/com/napnie/tfec/ui/RouteMapGUI.java
@@ -4,10 +4,6 @@ import com.napnie.tfec.PropertiesUtil;
 import com.sun.javafx.application.PlatformImpl;
 import java.awt.BorderLayout;
 import java.awt.Dimension;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-import javafx.application.Platform;
 import javafx.collections.ObservableList;
 import javafx.embed.swing.JFXPanel;
 import javafx.scene.Group;
@@ -16,50 +12,22 @@ import javafx.scene.Scene;
 import javafx.scene.web.WebEngine;
 import javafx.scene.web.WebView;
 import javafx.stage.Stage;
-import javax.swing.JButton;
 import javax.swing.JPanel;
 
 @SuppressWarnings("serial")
 public class RouteMapGUI extends JPanel {  
 
-	private JFXPanel jfxPanel;  
-	private JButton reload;  
+	private JFXPanel jfxPanel;
 	private WebEngine webEngine;  
 
 	private String[] mapPoint;
 
 	public RouteMapGUI(){
 		initComponents();
-		setPreferredSize(new Dimension(799, 626));
-	}  
-
-//	public static void main(String[] args){  
-//		RouteMapGUI ui = new RouteMapGUI();
-//		// Run this later:
-//		SwingUtilities.invokeLater(new Runnable() {  
-//			@Override
-//			public void run() {  
-//				final JFrame frame = new JFrame();  
-//
-//				frame.getContentPane().add( ui );  
-//
-//				frame.setMinimumSize(new Dimension(650, 450+75));  
-//				frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);  
-//				frame.setVisible(true);  
-//			}  
-//		});   
-//
-//		String prompt = "";
-//		Scanner scan = new Scanner(System.in);
-//		do {
-//			System.out.print("Origin ");
-//			String origin = scan.nextLine();
-//			System.out.print("Destination ");
-//			String destination = scan.nextLine();
-//			ui.setMap(origin, destination);
-//		} while( !prompt.equals("q") );
-//		scan.close();
-//	}  
+		Dimension dimension = new Dimension(799, 600 + 100);
+		setPreferredSize(dimension);
+		setMaximumSize(dimension);
+	}
 
 	private void initComponents(){  
 
@@ -67,25 +35,7 @@ public class RouteMapGUI extends JPanel {
 		createScene();  
 
 		setLayout(new BorderLayout());  
-		add(jfxPanel, BorderLayout.CENTER);  
-
-		reload = new JButton();  
-		reload.addActionListener(new ActionListener() {
-
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				Platform.runLater(new Runnable() {
-
-					@Override
-					public void run() {
-						webEngine.reload();
-					}
-				});
-			}
-		});  
-		reload.setText("Reload");  
-
-		add(reload, BorderLayout.SOUTH);  
+		add(jfxPanel, BorderLayout.CENTER);
 	}     
 
 	private void createScene() {  
@@ -122,7 +72,6 @@ public class RouteMapGUI extends JPanel {
 		mapPoint[0] = origin;
 		mapPoint[1] = destination;
 		createScene();
-		reload.doClick();
 	}
 	
 	private String getDefualtPlace() {
